Minimum image convention coded with a integar-based rounding for speed.
This appears to fix the PBC/Stencil errors.
See Allen and Tildesley p.24 onwards for description of various
applications of the algorithm.
In the printf log function, there is also an earlier version that
converts to floating point and uses the built-in round function.
This version *should* be much faster, albeit at the cost of complexity
at looking at it.
Not fully tested, but seems to mostly fix the issues with the edge of
the PBCs, at least on a quick desktop-based run.